Removed dupliate Unit enum
