+++ lib/models/period_unit.dart
@@ -1,0 +1,14 @@
+
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+enum PeriodUnit {
+  @JsonValue('DAY')
+  day,
+  @JsonValue('WEEK')
+  week,
+  @JsonValue('MONTH')
+  month,
+  @JsonValue('YEAR')
+  year,
+  unknown
+}
