+++ lib/models/period_wrapper.dart
@@ -1,5 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import 'period_unit.dart';
+
 part 'period_wrapper.freezed.dart';
 part 'period_wrapper.g.dart';
 
@@ -9,7 +11,7 @@ part 'period_wrapper.g.dart';
 class Period with _$Period {
   const factory Period(
     /// The number of period units: day, week, month, year, unknown
-    @JsonKey(name: 'unit') Unit unit,
+    @JsonKey(name: 'unit') PeriodUnit unit,
 
     /// The increment of time that a subscription period is specified in
     @JsonKey(name: 'value') int value,
@@ -21,17 +23,4 @@ class Period with _$Period {
   ) = _Period;
 
   factory Period.fromJson(Map<String, dynamic> json) => _$PeriodFromJson(json);
-}
-
-enum Unit {
-  @JsonValue('DAY')
-  day,
-  @JsonValue('WEEK')
-  week,
-  @JsonValue('MONTH')
-  month,
-  @JsonValue('YEAR')
-  year,
-  @JsonValue('UNKNOWN')
-  unknown,
 }
