+++ lib/object_wrappers.dart
@@ -7,6 +7,7 @@ export 'models/introductory_price.dart';
 export 'models/offering_wrapper.dart';
 export 'models/offerings_wrapper.dart';
 export 'models/package_wrapper.dart';
+export 'models/period_unit.dart';
 export 'models/period_wrapper.dart';
 export 'models/price_wrapper.dart';
 export 'models/pricing_phase_wrapper.dart';
